Rename purldb_project to purldb #821
