+++ docker-compose.common.yml
@@ -20,7 +20,7 @@ services:
     command: wait-for-it --strict --timeout=60 db:5432 -- sh -c "
         python manage_purldb.py migrate &&
         python manage_purldb.py collectstatic --no-input --verbosity 0 --clear &&
-        gunicorn purldb_project.wsgi:application --bind :8000 --timeout 600 --workers 8"
+        gunicorn purldb.wsgi:application --bind :8000 --timeout 600 --workers 8"
     env_file:
       - docker_purldb.env
     expose:
