+++ manage_purldb.py
@@ -14,5 +14,5 @@ import sys
 if __name__ == "__main__":
     from django.core.management import execute_from_command_line
 
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "purldb_project.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "purldb.settings")
     execute_from_command_line(sys.argv)
