+++ setup.cfg
@@ -102,7 +102,7 @@ dev =
 
 [options.entry_points]
 console_scripts =
-    purldb = purldb_project:command_line
+    purldb = purldb:command_line
 
 scancodeio_pipelines =
     matching = matchcode_pipeline.pipelines.matching:Matching
